moved cart and contact css w/ img to component folder
