+++ frontend/src/components/Contact/Contact.css
@@ -1,5 +1,5 @@
 .ContactScreen_hero-img{
-    background-image: url("../../../images/bg-graphics/screen-contact/contact-section-bg.png");
+    background-image: url("./images/contact-section-bg.png");
     background-repeat: no-repeat;
     background-attachment: fixed;
     height: 200px;
