+++ frontend/src/components/Contact/ContactHero.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import '../../style/2/contactScreen/Contact.css'
+import './Contact.css'
 import {Jumbotron, Container} from "reactstrap";
