+++ frontend/src/components/HeaderFooter/Footer.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {Row, Col,} from "react-bootstrap";
 import {LinkContainer} from "react-router-bootstrap";
 import logo from "../../images/bg-graphics/graphics/Final-Logo-Horizontal.png";
-import '../../style/Footer.css'
+import './Footer.css'
 import {Jumbotron} from "reactstrap";
 import SocialIcons from "../SocialIcons";
 import Fade from 'react-reveal';
